Ready Cursor Improvements
Updated tracking of initRoster positiong when a unit moves in updateUnitLocatin(). This was causing problems tracking where the ready cursor should be, and was a bug all round that would have bit me for various reasons.

Also updated clickTile logic to better display the ready cursor when a unit is not selected, and not display it when the ready unit IS selected.