Display Terrain Info Through AvailableCells
When selecting an available cell to move to, it is important to let players see the terrain info of those cells, so they can make informed decisions when they move. I updated the logic to display that info. I needed to create a new function or two, and re-arrange some logic with regard to FlashUnit() and DisplayTileInfo().

Introduced new global variable SelectedTile, which will function much like SelectedUnit.